+++ .env
@@ -46,7 +46,7 @@ OTEL_EXPORTER_OTLP_METRICS_TEMPORALITY_PREFERENCE=cumulative
 # Core Demo Services
 # ******************
 # Accounting Service
-ACCOUNTING_SERVICE_DOCKERFILE=./src/accountingservice/Dockerfile
+ACCOUNTING_DOCKERFILE=./src/accounting/Dockerfile
 
 # Ad Service
 AD_SERVICE_PORT=9555
