+++ .licenserc.json
@@ -38,7 +38,6 @@
     ],
     "ignore": [
         "node_modules/",
-        "src/accountingservice/genproto/",
         "src/cartservice/src/obj/",
         "src/cartservice/tests/obj/",
         "src/checkoutservice/genproto/",
