+++ ide-gen-proto.sh
@@ -65,7 +65,7 @@ gen_proto_ts() {
   cd "$base_dir" || return
 }
 
-gen_proto_dotnet accountingservice
+gen_proto_dotnet accounting
 # gen_proto_java adservice
 gen_proto_dotnet cartservice
 gen_proto_go checkoutservice
