+++ src/accounting/Consumer.cs
@@ -5,7 +5,7 @@ using Confluent.Kafka;
 using Microsoft.Extensions.Logging;
 using Oteldemo;
 
-namespace AccountingService;
+namespace Accounting;
 
 internal class Consumer : IDisposable
 {
@@ -74,7 +74,7 @@ internal class Consumer : IDisposable
     {
         var conf = new ConsumerConfig
         {
-            GroupId = $"accountingservice",
+            GroupId = $"accounting",
             BootstrapServers = servers,
             // https://github.com/confluentinc/confluent-kafka-dotnet/tree/07de95ed647af80a0db39ce6a8891a630423b952#basic-consumer-example
             AutoOffsetReset = AutoOffsetReset.Earliest,
